cassandra-gateway-server/src/requirements.txt
tornado

cassandra-driver
lz4

